+++ wasmdclient/client/log.go
