+++ wasmdclient/config/wasmd_query_config_test.go
