+++ wasmdclient/query/client.go
